moved script right before closing body tag
I moved ‘pagination.js’ right before the closing body tag. It was
before the last closing div.

I also added the pagination script and jquery to the other student
lists to ensure it works with any length of list.

Tested in Firefox, Chrome, and Safari. Responsiveness looks good as
well.